Adding read_temperature
Adding a getter that reads the temperature, based on numbers from the datasheet. The 21's happen to be the same but are separately defined in the formula, so should be separate constants. 

It is a read, and not a property since it does have to fetch registers from the device.